aider: Moved the `ondataavailable` event handler inside the click event listener for the 'start' button in order to fix the error.
# Aider chat conversation:

USER: Uncaught TypeError: Cannot set properties of undefined (setting 'ondataavailable')

ASSISTANT: The error message "Uncaught TypeError: Cannot set properties of undefined (setting 'ondataavailable')" indicates that the `mediaRecorder` object is not defined at the time when the `ondataavailable` event handler is being set. This is because the `mediaRecorder` object is defined inside the click event listener for the 'start' button, but the `ondataavailable` event handler is being set outside of this event listener.

To fix this issue, we need to move the `ondataavailable` event handler inside the click event listener for the 'start' button, after the `mediaRecorder` object is defined.

Here are the changes to be made in the `script.js` file:

```javascript
script.js
<<<<<<< HEAD
mediaRecorder.ondataavailable = (event) => {
    if (event.data.size > 0) {
        recordedBlobs.push(event.data);
        const audioBlob = new Blob(recordedBlobs, {type: 'audio/webm'});
        const audioUrl = URL.createObjectURL(audioBlob);
        const audio = document.querySelector('#player');
        audio.src = audioUrl;
    }
};
=======
document.querySelector('#start').addEventListener('click', async () => {
    const stream = await navigator.mediaDevices.getUserMedia({audio: true});
    mediaRecorder = new MediaRecorder(stream);
    recordedBlobs = [];
    mediaRecorder.ondataavailable = (event) => {
        if (event.data.size > 0) {
            recordedBlobs.push(event.data);
        }
    };
    mediaRecorder.start();
    document.querySelector('#start').disabled = true;
    document.querySelector('#stop').disabled = false;

    mediaRecorder.ondataavailable = (event) => {
        if (event.data.size > 0) {
            recordedBlobs.push(event.data);
            const audioBlob = new Blob(recordedBlobs, {type: 'audio/webm'});
            const audioUrl = URL.createObjectURL(audioBlob);
            const audio = document.querySelector('#player');
            audio.src = audioUrl;
        }
    };
});
>>>>>>> updated
```

This change will ensure that the `ondataavailable` event handler is set after the `mediaRecorder` object is defined, which should resolve the error.